Discussion on HTTP/2
Please comment on:

Note that we don't recommend enabling query batching with Engine for most uses. Engine supports HTTP/2 even when your server doesn't, and one of the features that allows is sending multiple requests in parallel over a single connection. In most cases, that is actually more efficient than query batching, because requests are not blocked on each other. If you want to use HTTP/2, make sure to use the sidecar in single proxy mode or deploy Engine using the Docker container.